Split Platform primitives from networking
Platforms using LwIP needs to use the "linux" implementation of networking.

Not every platform using LwIP has support for every linux primitives/syscalls (such as signal).

To allow the adoption in these kind of devices split the network from OS primitives.
