+++ MQTTClient-C/src/linux/MQTTLinuxNetwork.c
@@ -17,49 +17,8 @@
 
 #include "MQTTLinux.h"
 
-void TimerInit(Timer* timer)
-{
-	timer->end_time = (struct timeval){0, 0};
-}
 
-char TimerIsExpired(Timer* timer)
-{
-	struct timeval now, res;
-	gettimeofday(&now, NULL);
-	timersub(&timer->end_time, &now, &res);
-	return res.tv_sec < 0 || (res.tv_sec == 0 && res.tv_usec <= 0);
-}
-
-
-void TimerCountdownMS(Timer* timer, unsigned int timeout)
-{
-	struct timeval now;
-	gettimeofday(&now, NULL);
-	struct timeval interval = {timeout / 1000, (timeout % 1000) * 1000};
-	timeradd(&now, &interval, &timer->end_time);
-}
-
-
-void TimerCountdown(Timer* timer, unsigned int timeout)
-{
-	struct timeval now;
-	gettimeofday(&now, NULL);
-	struct timeval interval = {timeout, 0};
-	timeradd(&now, &interval, &timer->end_time);
-}
-
-
-int TimerLeftMS(Timer* timer)
-{
-	struct timeval now, res;
-	gettimeofday(&now, NULL);
-	timersub(&timer->end_time, &now, &res);
-	//printf("left %d ms\n", (res.tv_sec < 0) ? 0 : res.tv_sec * 1000 + res.tv_usec / 1000);
-	return (res.tv_sec < 0) ? 0 : res.tv_sec * 1000 + res.tv_usec / 1000;
-}
-
-
-int linux_read(Network* n, unsigned char* buffer, int len, int timeout_ms)
+static int linux_read(Network* n, unsigned char* buffer, int len, int timeout_ms)
 {
 	struct timeval interval = {timeout_ms / 1000, (timeout_ms % 1000) * 1000};
 	if (interval.tv_sec < 0 || (interval.tv_sec == 0 && interval.tv_usec <= 0))
@@ -92,7 +51,7 @@ int linux_read(Network* n, unsigned char* buffer, int len, int timeout_ms)
 }
 
 
-int linux_write(Network* n, unsigned char* buffer, int len, int timeout_ms)
+static int linux_write(Network* n, unsigned char* buffer, int len, int timeout_ms)
 {
 	struct timeval tv;
